// types/types.ts
export interface MediaItem {
    title: string
    artist: string
    price: string
    imageUrl: string
    endTime: string
  }
  
  export interface Release {
    title: string
    artist: string
    duration: string
    timeAgo: string
    imageUrl: string
  }
  
  export interface Video {
    title: string
    thumbnailUrl: string
    duration: string
  }